backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
yfinance==0.2.32
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
python-dateutil==2.8.2